pyproject.toml
[project]
name = "snackbase"
version = "0.1.0"
description = "Open-source Backend-as-a-Service (BaaS) - Self-hosted alternative to PocketBase"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.127.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.30.0",
    "greenlet>=3.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    "structlog>=24.4.0",
    "python-multipart>=0.0.12",
    "click>=8.1.0",
    "argon2-cffi>=23.1.0",
    "pyjwt>=2.8.0",
    "email-validator>=2.1.0",
    "alembic>=1.17.2",
    "cryptography>=46.0.3",
    "httpx>=0.28.1",
    "signxml>=4.2.0",
    "lxml>=6.0.2",
]

[project.scripts]
snackbase = "snackbase.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/snackbase"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "jinja2>=3.1.6",
    "respx>=0.22.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["snackbase"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
markers = [
    "enable_audit_hooks: Enable audit logging hooks for this test (adds overhead)",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

